+++ packages/model-viewer/src/test/features/staging-spec.ts
@@ -65,12 +65,11 @@ suite('Staging', () => {
         expect(element.turntableRotation).to.be.greaterThan(turntableRotation);
       });
 
-      // TODO(#1205)
-      test.skip(
+      test(
           'retains turntable rotation when auto-rotate is toggled',
           async () => {
             element.autoRotateDelay = 0;
-            await timePasses();
+            await rafPasses();
             await rafPasses();
 
             const {turntableRotation} = element;
@@ -78,23 +77,23 @@ suite('Staging', () => {
             expect(turntableRotation).to.be.greaterThan(0);
 
             element.autoRotate = false;
-            await timePasses();
+            await rafPasses();
             await rafPasses();
 
             expect(element.turntableRotation).to.be.equal(turntableRotation);
 
             element.autoRotate = true;
-            await timePasses();
+            await rafPasses();
             await rafPasses();
 
             expect(element.turntableRotation)
                 .to.be.greaterThan(turntableRotation);
           });
 
-      // TODO(#1206)
-      test.skip('pauses rotate after user interaction', async () => {
+      test('pauses rotate after user interaction', async () => {
         const {turntableRotation} = element;
         await timePasses(AUTO_ROTATE_DELAY);
+        await rafPasses();
         await rafPasses();
 
         const {turntableRotation: initialTurntableRotation} = element;
@@ -103,13 +102,13 @@ suite('Staging', () => {
         element.dispatchEvent(new CustomEvent<CameraChangeDetails>(
             'camera-change',
             {detail: {source: ChangeSource.USER_INTERACTION}}));
-        await timePasses();
-
+        await rafPasses();
         await rafPasses();
 
         expect(element.turntableRotation).to.be.equal(initialTurntableRotation);
 
         await timePasses(AUTO_ROTATE_DELAY);
+        await rafPasses();
         await rafPasses();
 
         expect(element.turntableRotation)
